Module delegator slashing protection
Module delegator slashing protection. In some cases, we need to delegate using the module and need a feature to protect such delegation from the validator slashing. The implementation contains the automatic withdrawal reward and undelegation just before the validator slashing. The slashing amount is computed based on the updated power after the undelegation.

Don't fail the in case the UnbondAndUndelegateCoins ammount is zero